update api and web admin

+++ API/src/main/java/com/example/api/Service/UserServiceImpl.java
@@ -6,10 +6,13 @@ import com.example.api.Entity.User;
 //import com.example.api.Entity.User_Role;
 import com.example.api.Exception.AppException;
 
+import com.example.api.Exception.UserAlreadyExistException;
+import com.example.api.Exception.UserNotFoundException;
 import com.example.api.Repository.RoleRepository;
 import com.example.api.Repository.UserRepository;
 //import com.example.api.Repository.User_RoleRepository;
 
+import com.example.api.Service.IService.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Service;
@@ -26,9 +29,18 @@ public class UserServiceImpl implements UserService {
     private RoleRepository roleRepository;
 
 
+    @Override
+    public List<String> getAllMarjor() {
+        return userRepository.getAllMarjor();
+    }
 
-//    @Autowired
-//    User_RoleRepository user_roleRepository;
+
+
+    public List<User> findByMajorWithRole(@Param("major") String major)
+    {
+        return userRepository.findByMajorWithRole(major);
+    }
+
 
     @Override
     public List<User> getAllUsers() {
@@ -46,9 +58,9 @@ public class UserServiceImpl implements UserService {
 
         Optional<User> checkUser = userRepository.findByUsername(user.getUsername());
         if(checkUser.isPresent()){
-//            throw new UserAlreadyExistException(user.getUsername() + "User already exist");
+            throw new UserAlreadyExistException(user.getUsername() + "User already exist");
         }
-        Role role =roleRepository.findByName("STUDENT").get();
+        Role role =roleRepository.findByName("Student").get();
         user.setRoles(Collections.singletonList(role));
         return userRepository.save(user);
 
@@ -57,38 +69,66 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateUser(UUID userId, User user) {
-        // Check if the user exists
-        if (userRepository.existsById(userId)) {
-            user.setId(userId);
-            return userRepository.save(user);
+        User existingUser = userRepository.findById(userId).orElse(null);
+        if (existingUser != null) {
+            if(user.getUsername() != null){
+                existingUser.setUsername(user.getUsername());
+            }
+            if(user.getPassword() != null){
+                existingUser.setPassword(user.getPassword());
+            }
+            if(user.getEmail() != null){
+                existingUser.setEmail(user.getEmail());
+            }
+            if(user.getFullname() != null){
+                existingUser.setFullname(user.getFullname());
+            }
+            if(user.getMajor() != null){
+                existingUser.setMajor(user.getMajor());
+            }
+            if(user.getAvatar() != null){
+                existingUser.setAvatar(user.getAvatar());
+            }
+            if(user.getMssv() != null){
+                existingUser.setMssv(user.getMssv());
+            }
+
+            return userRepository.save(existingUser);
         }
-        return null; // or throw an exception indicating that the user does not exist
+        throw new UserNotFoundException("Không tìm thấy người dùng với Id: " + userId);
     }
 
     @Override
     public void deleteUser(UUID userId) {
-        userRepository.deleteById(userId);
+        Optional<User> optionalUser = userRepository.findById(userId);
+
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(userId);
+        }else {
+            throw  new UserNotFoundException("Không tìm thấy người dùng với Id: " + userId);
+        }
     }
 
+    @Override
+    public User findByName(String name) {
+        return userRepository.findByUsername(name).get();
+    }
+
+    @Override
+    public Collection<Role> getUserRoles(UUID userId) {
+        User user = userRepository.findById(userId).orElse(null);
+        if (user != null) {
+            return user.getRoles(); // Assuming there's a method in User entity to get roles
+        }
+        return Collections.emptyList();
+    }
     @Override
     public List<User> findByMajor(String major) {
-//        List<User> users = userRepository.findByMajor(major);
-//        if(users.isEmpty()){
-////            throw new UserNotFoundException("No users found with major: " + major);
-//        }
-      return null;
-    }
-
-    @Override
-    public List<String> getAllMarjor() {
-        return userRepository.getAllMarjor();
-    }
-
-
-
-    public List<User> findByMajorWithRole(@Param("major") String major)
-    {
-        return userRepository.findByMajorWithRole(major);
+        List<User> users = userRepository.findByMajor(major);
+        if(users.isEmpty()){
+            throw new UserNotFoundException("No users found with major: " + major);
+        }
+        return users;
     }
 
 }
